pyproject.toml
[tool.poetry]
name = "airbyte-deployment"
version = "0.1.0"
description = ""
authors = ["donatien konan <donatien.konan.pro@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pendulum = "*"
pytest = "*"
apache-airflow = {version = "^2.4.0", extras = ["postgres", "airbyte"]}
dbt-core = "*"
pyarrow = "*"
pandas = "1.5.3"
pytest-mock = "^3.14.0"
dbt-clickhouse = "1.8.4"
apache-airflow-providers-amazon = "8.23.0"
minio = "^7.2.15" # MinIO Python SDK
shapely = "2.0.6"

[tool.poetry.group.dev.dependencies]
ipython = "8.8.0"
missingno = "0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
